Completion $ appears in the center of the blue progress bar
This is more logical than the proposed UI, because if the % appears on the white portion, the user may get confused as to what the number refers to (% done or incomplete). The text is now centered in the blue portion. Also, fixed a small bug where 100% completion does not fill up the blue progress bar.